+++ app/build.gradle.kts
@@ -70,9 +70,6 @@ dependencies {
     implementation(Dependencies.Firebase.KTX)
 
     implementation(Dependencies.KAKAO_LOGIN)
-    implementation("androidx.appcompat:appcompat:1.3.0")
-    implementation("org.jetbrains.kotlin:kotlin-stdlib:${rootProject.extra["kotlin_version"]}")
-    implementation("androidx.constraintlayout:constraintlayout:2.0.4")
 
     testImplementation(Dependencies.Test.JUNIT)
     androidTestImplementation (Dependencies.AndroidXTest.JUNIT)
