+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ExchangeDiary"
         android:usesCleartextTraffic="true">
-        <activity android:name=".ui.select_next_writer.SelectNextWriterActivity"></activity>
         <activity android:name=".ui.bridge.BridgeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,6 +23,7 @@
         <activity android:name=".ui.home.HomeActivity" />
         <activity android:name=".ui.diary.DiaryActivity" />
         <activity android:name=".ui.notification_page.NotificationPageActivity" />
+        <activity android:name=".ui.select_next_writer.SelectNextWriterActivity"/>
 
         <meta-data
             android:name="com.kakao.sdk.AppKey"
